Make the positive label an independent component to not add too much complexity to the Target Variable component.
